final-qemie合并进入dynamic-link分支

See merge request T202510486995232/oskernel2025!9

+++ kernel/pm/process_manager.cc
@@ -2316,10 +2316,6 @@ namespace pm
 		else // normal file
 		{
 			fs::normal_file *f = new fs::normal_file( attrs, dentry );
-			if ( dentry->rName() == "test.txt" ) {
-				// printf( "dup success\n" );
-				dentry->dup();
-			}
 			// log_info( "test normal file read" );
 			// {
 			// 	fs::file *ff = ( fs::file * ) f;
@@ -2614,6 +2610,7 @@ namespace pm
 				return -1;
 
 			if ( path[0] == '.' && path[1] == '/' ) path = path.substr( 2 );
+			if ( path == "/proc/interrupts" ) return -1;
 
 			return fs::k_file_table.unlink( path );
 		}
